+++ champss/pipeline_batch_db/sps_pipeline/pipeline.py
@@ -46,7 +46,7 @@ from sps_pipeline import (  # ps,
     utils,
 )
 
-datpath = "/data/chime/sps/raw"
+default_datpath = "/data/chime/sps/raw"
 
 
 def load_config(config_file="sps_config.yml"):
@@ -294,6 +294,17 @@ def dbexcepthook(type, value, tb):
     type=float,
     help="Frequency at which to stop processing candidates.",
 )
+@click.option(
+    "--scale-injections/--not-scale-injections",
+    default=False,
+    help="Scale injection so that input sigma should be detected sigma.",
+)
+@click.option(
+    "--datpath",
+    default=default_datpath,
+    type=str,
+    help="Path to the raw data folder.",
+)
 def main(
     date,
     stack,
@@ -318,6 +329,8 @@ def main(
     injection_idx,
     only_injections,
     cutoff_frequency,
+    scale_injections,
+    datpath,
 ):
     """
     Runner script for the Slow Pulsar Search prototype pipeline v0.
@@ -439,19 +452,6 @@ def main(
             active_pointings = strat.active_pointing_from_pointing(
                 closest_pointing, date
             )
-            for active_pointing in active_pointings:
-                data_list.extend(
-                    get_data_list(
-                        active_pointing.max_beams, basepath=datpath, extn="dat"
-                    )
-                )
-            if not data_list:
-                log.error(
-                    "No data found for the pointing {:.2f} {:.2f}".format(
-                        active_pointings[0].ra, active_pointings[0].dec
-                    )
-                )
-                sys.exit()
             for obs_id_file in obs_id_files:
                 with open(obs_id_file) as infile:
                     sub_pointing = int(obs_id_file.split("obs")[0].split("_")[-2])
@@ -468,20 +468,6 @@ def main(
             active_pointings = strat.active_pointing_from_pointing(
                 closest_pointing, date
             )
-            data_list = []
-            for active_pointing in active_pointings:
-                data_list.extend(
-                    get_data_list(
-                        active_pointing.max_beams, basepath=datpath, extn="dat"
-                    )
-                )
-            if not data_list:
-                log.error(
-                    "No data found for the pointing {:.2f} {:.2f}".format(
-                        active_pointings[0].ra, active_pointings[0].dec
-                    )
-                )
-                sys.exit()
             # Record the obs id
             for active_pointing in active_pointings:
                 obs_id_file = os.path.join(
@@ -567,7 +553,9 @@ def main(
                 )
                 fdmt = False
             if "beamform" in components:
-                beamformer = beamform.initialise(config, rfi_beamform, basepath)
+                beamformer = beamform.initialise(
+                    config, rfi_beamform, basepath, datpath
+                )
                 skybeam, spectra_shared = beamform.run(
                     active_pointing, beamformer, fdmt, num_threads, basepath
                 )
@@ -624,6 +612,7 @@ def main(
                     **OmegaConf.to_container(config.ps),
                     run_ps_stack=stack,
                     num_threads=num_threads,
+                    known_source_threshold=known_source_threshold,
                 )
                 # Use global to allow unlinking memory in exception hook
                 global power_spectra
@@ -643,6 +632,7 @@ def main(
                         injection_idx,
                         only_injections,
                         cutoff_frequency,
+                        scale_injections,
                         obs_folder,
                         prefix,
                     )
@@ -871,6 +861,11 @@ def main(
     default=100,
     help="Frequency at which to stop processing candidates.",
 )
+@click.option(
+    "--scale-injections/--not-scale-injections",
+    default=False,
+    help="Scale injection so that input sigma should be detected sigma.",
+)
 def stack_and_search(
     plot,
     plot_threshold,
@@ -887,6 +882,7 @@ def stack_and_search(
     injection_idx,
     only_injections,
     cutoff_frequency,
+    scale_injections,
 ):
     """
     Runner script to stack monthly PS into cumulative PS and search the eventual stack.
@@ -957,6 +953,7 @@ def stack_and_search(
             injection_idx,
             only_injections,
             cutoff_frequency,
+            scale_injections,
         )
         ps_stack = db_api.get_ps_stack(closest_pointing._id)
         if ps_detections_monthly:
@@ -970,6 +967,10 @@ def stack_and_search(
                 plot,
                 plot_threshold,
                 config.cands.get("write_harmonically_related_clusters", False),
+                False,
+                injection_path,
+                injection_idx,
+                only_injections,
             )
     else:
         power_spectra_monthly = None
@@ -983,6 +984,7 @@ def stack_and_search(
             injection_idx,
             only_injections,
             cutoff_frequency,
+            scale_injections,
         )
         if to_search:
             if not ps_detections:
@@ -1004,6 +1006,10 @@ def stack_and_search(
                         plot,
                         plot_threshold,
                         config.cands.get("write_harmonically_related_clusters", False),
+                        False,
+                        injection_path,
+                        injection_idx,
+                        only_injections,
                     )
     try:
         power_spectra_monthly.unlink_shared_memory()
@@ -1066,7 +1072,15 @@ def stack_and_search(
     type=str,
     help="Name used for the mongodb database.",
 )
-def find_pointing_with_data(date, beam, full_transit, db_port, db_host, db_name):
+@click.option(
+    "--datpath",
+    default=default_datpath,
+    type=str,
+    help="Path to the raw data folder.",
+)
+def find_pointing_with_data(
+    date, beam, full_transit, db_port, db_host, db_name, datpath
+):
     """
     Script to determine the pointings with data to process in a given day.